--- Assets/Project/Commons/DataBase/Scripts/DataBase.asmdef
@@ -1,16 +1,0 @@
-{
-    "name": "DataBase",
-    "rootNamespace": "",
-    "references": [
-        "GUID:9160a9816a6c1104395215dc6d88ddcf"
-    ],
-    "includePlatforms": [],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": [],
-    "versionDefines": [],
-    "noEngineReferences": false
-}
