+++ Assets/Project/Scenes/Title/Scripts/Presenter/TitleScenePresenter.cs
@@ -1,4 +1,4 @@
-using Project.Commons.Scripts.Model;
+using Project.Scripts.Model;
 using Project.Scenes.Title.Scripts.Model;
 using Project.Scenes.Title.Scripts.View;
 using UniRx;
@@ -12,13 +12,13 @@ namespace Project.Scenes.Title.Scripts.Presenter
         [SerializeField] TitleMenuView titleMenuView;
         [SerializeField] TitleSettingModalView titleSettingModalView;
         
-        TitleBackgroundModel _titleBackgroundModel;
-        TitleGameStartModel _titleGameStartModel;
+        TitleBackgroundModel titleBackgroundModel;
+        TitleGameStartModel titleGameStartModel;
 
         void Start()
         {
-            _titleBackgroundModel = new TitleBackgroundModel();
-            _titleGameStartModel = new TitleGameStartModel();
+            titleBackgroundModel = new TitleBackgroundModel();
+            titleGameStartModel = new TitleGameStartModel();
             
             titleMenuView.OnPressedStart.Subscribe(StartGame);
             titleMenuView.OnPressedExit.Subscribe(ExitGame);
@@ -30,7 +30,7 @@ namespace Project.Scenes.Title.Scripts.Presenter
 
         void StartGame(Unit _)
         {
-            _titleGameStartModel.StartGame();
+            titleGameStartModel.StartGame();
             SceneManager.LoadScene(SceneRouterModel.QuestList, LoadSceneMode.Additive);
             SceneManager.SetActiveScene(SceneManager.GetSceneByName(SceneRouterModel.QuestList));
             SceneManager.UnloadSceneAsync(gameObject.scene.name);
@@ -47,7 +47,7 @@ namespace Project.Scenes.Title.Scripts.Presenter
 
         void SetTitleBackGround()
         {
-            var clearedStageAmount = _titleBackgroundModel.ClearedStageAmount;
+            var clearedStageAmount = titleBackgroundModel.ClearedStageAmount;
             titleMenuView.SetBackGround(clearedStageAmount);
         }
     }
