+++ Assets/Project/Scripts.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 9160a9816a6c1104395215dc6d88ddcf
-AssemblyDefinitionImporter:
+guid: 6e41947f988a4d34eab0640181ae12cd
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
