+++ Assets/Project/Scripts/Extensions/BoolExtensions.cs
