+++ Assets/Project/Scripts/Extensions/Extensions.asmdef
@@ -1,0 +1,3 @@
+{
+	"name": "Extensions"
+}
