+++ Assets/Project/Scripts/Extensions/StringExtensions.cs
