+++ Assets/Project/Scripts/Infra/Infra.asmdef
@@ -1,11 +1,8 @@
 {
-    "name": "Extensions",
+    "name": "Infra",
     "rootNamespace": "",
     "references": [
-        "GUID:f51ebe6a0ceec4240a699833d6309b23",
-        "GUID:029c1c1b674aaae47a6841a0b89ad80e",
-        "GUID:560b04d1a97f54a4e82edc0cbbb69285",
-        "GUID:1a6be227346ac8745a511432f2258344"
+        "GUID:dd4687c0631a8344488f851ae4d2049b"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
