+++ Assets/Project/Scripts/Infra/Infra.asmdef.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: e94caeccb95cdb74aa26cb2e5542e358
-folderAsset: yes
-DefaultImporter:
+guid: 1d039aa9328089146b1df536a796f39e
+AssemblyDefinitionImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
