+++ Assets/Project/Scripts/Model/Model.asmdef
@@ -6,9 +6,9 @@
         "GUID:84651a3751eca9349aac36a66bba901b",
         "GUID:560b04d1a97f54a4e82edc0cbbb69285",
         "GUID:f51ebe6a0ceec4240a699833d6309b23",
-        "GUID:9160a9816a6c1104395215dc6d88ddcf",
         "GUID:1ac93fd9520186a4eaf0cd7243e8a18c",
-        "GUID:a21533655bbf1384e965b16ef1916e6b"
+        "GUID:dd4687c0631a8344488f851ae4d2049b",
+        "GUID:1d039aa9328089146b1df536a796f39e"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
