+++ Assets/Project/Textures/CharaImage/Ushitora_Still.png.meta
@@ -186,7 +186,8 @@ TextureImporter:
     secondaryTextures: []
     spriteCustomMetadata:
       entries: []
-    nameFileIdTable: {}
+    nameFileIdTable:
+      Ushitora_Still_0: 669692194207701178
   mipmapLimitGroupName: 
   pSDRemoveMatte: 0
   userData:
