--- Assets/QuestListScenePresenter.cs
@@ -1,8 +1,0 @@
-using Project.Scenes.QuestList.Scripts.Model;
-using UnityEngine;
-
-public class QuestListScenePresenter : MonoBehaviour
-{
-    readonly QuestModel questModel = new();
-
-}
