--- Assets/QuestListScenePresenter.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 0410b597cf874a948a568a91e833e839
